# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Penjualan",
        "description": "API untuk manajemen data penjualan",
        "contact": {
            "name": "Muhammad Gibran",
            "email": "gibran.devv@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/detail-penjualan/{id}": {
            "get": {
                "tags": [
                    "Detail Penjualan"
                ],
                "summary": "Mendapatkan detail penjualan berdasarkan ID",
                "operationId": "dcd781fbdefee0071e4598446b2f747f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan detail penjualan"
                    },
                    "404": {
                        "description": "Detail penjualan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Detail Penjualan"
                ],
                "summary": "Memperbarui data detail penjualan",
                "operationId": "1c8997092f9f3276f5b498065834eec9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID detail penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data detail penjualan yang diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_penjualan": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_produk": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "example": 150000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data detail penjualan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Detail Penjualan"
                ],
                "summary": "Menghapus detail penjualan",
                "operationId": "5dc0a7624b34c1810d90e1f77916f969",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID detail penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus detail penjualan"
                    },
                    "404": {
                        "description": "Detail penjualan tidak ditemukan"
                    }
                }
            }
        },
        "/api/detail-penjualan/create": {
            "post": {
                "tags": [
                    "Detail Penjualan"
                ],
                "summary": "Membuat detail penjualan baru",
                "operationId": "040991d9fe957b0836d287aa58e47d42",
                "requestBody": {
                    "description": "Data detail penjualan baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_penjualan": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_produk": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat detail penjualan baru"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
    },
...
}